Add rust-geo dependency
